fix: レビュー指摘対応（Step.positionをi16化、started_at/completed_atをJSTに統一、start()へリネーム、remove_parameterテスト修正） in response to review feedback

+++ backend/src/repository/trial_repo.rs
@@ -203,7 +203,7 @@ impl TrialRepository for PgTrialRepository {
                     .bind(step.id().0)
                     .bind(trial.id().0)
                     .bind(step.name())
-                    .bind(step.position() as i16)
+                    .bind(step.position())
                     .bind(step.started_at().copied())
                     .bind(step.completed_at().copied()),
                 )
